pyproject.toml
[project]
name = "sharing-hub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bootstrap4>=0.1.0",
    "bs4>=0.0.2",
    "celery>=5.6.2",
    "django==4.2.*",
    "django-admin-honeypot-2024>=1.0.0",
    "django-bootstrap4>=26.1",
    "django-celery-beat>=2.8.1",
    "django-cookie-law>=2.2.0",
    "django-debug-toolbar-template-timings>=0.9",
    "django-haystack>=3.3.0",
    "django-memcache-status>=2.3",
    "django-otp>=1.7.0",
    "django-redis>=6.0.0",
    "django-simple-history>=3.11.0",
    "django-utils-six>=2.0",
    "flower>=2.0.1",
    "haystack>=0.42",
    "lazy-object-proxy>=1.12.0",
    "pillow>=12.1.0",
    "psycopg2>=2.9.11",
    "pymemcache>=4.0.0",
    "python-memcached>=1.62",
    "redis>=7.1.0",
    "sorl-thumbnail>=13.0.0",
    "zeep>=4.3.2",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.2.0",
]
